% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-algorithm_settings.R
\name{algorithm_settings}
\alias{algorithm_settings}
\title{Specify Algorithm Settings for 'brm'}
\usage{
algorithm_settings(
  desired_iter = 10000,
  rng_seed = NULL,
  warmup = 500,
  adapt_delta = 0.9,
  max_treedepth = 15,
  cores = 4,
  chains = 4,
  alpha = 0.05,
  quick_fit = F
)
}
\arguments{
\item{desired_iter}{The total number of samples to draw
from the posterior (the sum of individual chain).}

\item{rng_seed}{RNG seed used to ensure reproducibility.}

\item{warmup}{The number of iterations to use for the
warm-up period before drawing samples from the posterior.}

\item{adapt_delta}{Proportion governing the gradient descent
algorithm; higher values result in slower but more
stable estimates (i.e., lower risk of divergent estimates
where gradient descent overshoots).}

\item{max_treedepth}{Integer governing steps used by
gradient descent; higher values result in slower but
more stable estimation.}

\item{cores}{Number of cores to use for parallel processing.}

\item{chains}{Number of chains to estimate (higher number
provides better estimate of convergence diagnostics).}

\item{alpha}{Cut-off for significance (will be adjusted for
Monte Carlo).}

\item{quick_fit}{Logical; if \code{TRUE} adjusts settings
for a quicker, less stable fit (good for initial model
specification).}
}
\value{
A list with the algorithm settings.
\enumerate{
\item total: The total number of posterior samples to draw.
\item warmup: The number of iterations to use for the warm-up
period before sampling from the posterior.
\item iter: The number of iterations to use per chain, the sum
of the warm-up iterations and the division of the total samples
by the number of chains.
\item adapth_delta: A proportion governing the gradient descent
settings.
\item max_treedepth: An integer governing the gradient descent
settings.
\item cut_off: The cut-off for statistical significant after
correcting for Monte Carlo error (by taking the lower
boundary of the 95\% confidence interval around the
specified alpha).
}
}
\description{
A convenience function used to specify settings
when sampling from the joint posterior of a model
fit using \code{\link[brms]{brm}}. Can be used to
set the number of samples from the posterior to draw,
the warm-up period used pre-sampling, the number
of chains and cores, and additional control options.
}
\examples{
# Default
alg <- algorithm_settings()

# Settings for faster, less stable model fitting
alg <- algorithm_settings( desired_iter = 2000,
                           adapt_delta = .8,
                           max_treedepth = 12 )

}
