% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions_for_estimation.R
\name{ci}
\alias{ci}
\title{Compute Credible Intervals From Monte Carlo Samples}
\usage{
ci(x, width = 0.95, index = 1:2)
}
\arguments{
\item{x}{A vector of values, Monte Carlo samples for
a prior or posterior distribution.}

\item{width}{The desired width of the credible interval.}

\item{index}{The index for the bounds to return, either
\code{1} for the lower bound, \code{2} for the upper
bound, for the vector \code{1:2} for both (default).}
}
\value{
A vector giving the lower and/or upper bound
for the credible interval of the specified width.
}
\description{
Function to compute credible intervals of a desired
width from a vector of Monte Carlo samples via
the base R \code{\link[stats]{quantile}} function.
}
\examples{
set.seed( 831 ) # For reproducibility
x <- rnorm( 1000, mean = 1 )
# 95\% CI
ci( x )
# Approx. one standard deviation from
# the mean (interval is close to [0,1])
ci( x, width = .68 )
# Lower or upper bound
ci( x, .68, 1 )
ci( x, .68, 2 )

}
