% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extbrms-package.R
\docType{package}
\name{extbrms-package}
\alias{extbrms-package}
\alias{extbrms}
\title{Convenience Functions for the 'brms' Package}
\description{
The \pkg{extbrms} package implements a set of
convenience functions that interface with the
package \pkg{brms} to aid in specifying algorithm
settings, saving output, plotting results, and
computing contrasts. A complete  list of the functions
available in the package can be obtained via the command
\code{ls(pos = "package:extbrms")}.
}
