% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions_for_estimation.R
\name{\%s\%}
\alias{\%s\%}
\title{Operator to Select Parameters}
\usage{
x \%s\% y
}
\arguments{
\item{x}{Either a \code{\link[brms]{brmsfit}}
object,a matrix of posterior samples, a
character vector, or a list with an
element \code{fit} that is a
\code{\link[brms]{brmsfit}} object.}

\item{y}{A character string, to be
converted into an R expression used
to match and select parameter names.}
}
\value{
A character vector.
}
\description{
An operator that takes an input
(such as a \code{\link[brms]{brmsfit}}
object, matrix, or output from the
\code{\link{fit_with_brm}} function)
and extracts a character vector with
the parameter names that meet a
specified set of criteria.
}
\details{
The operator converts the character
string on the right-hand side into an
expression making calls to
\code{\link[base]{grepl}} and
\code{&}, \code{|}, and \code{!}.
}
\examples{
# Example vector of parameter names
x <- c(
  'b_Intercept', 'b_speed', 'sigma',
  'prior_Intercept', 'prior_sigma', 'lp__'
)

# Extract all parameters
x \%s\% ''

# Extract parameters with 'b_'
x \%s\% 'b_'

# Extract parameters with both 'b_' and 'speed'
x \%s\% 'b_ & speed'

# Extract parameters with 'b_' but not 'speed'
x \%s\% 'b_ & - speed'

# Extract parameters with 'b_' or 'sigma'
x \%s\% 'b_ | sigma'

# Combinations of '&' and '|'
x \%s\% '( b_ | sigma ) & - prior'

}
