% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions_for_estimation.R
\name{p_value.monte_carlo}
\alias{p_value.monte_carlo}
\title{Compute P-values From Monte Carlo Samples}
\usage{
p_value.monte_carlo(x, alternative = "two-sided", comparison = 0)
}
\arguments{
\item{x}{A vector of values.}

\item{alternative}{A character string specifying the
alternative hypothesis, either \code{'two-sided'}
(default), \code{'greater'}, or \code{'less'}.
You can specify just the initial letter.}

\item{comparison}{A cut-off value; the function
computes the proportion of values in \code{x} that fall
above or below \code{comparison}.}
}
\value{
A proportion between 0 and 1.
}
\description{
A general-purpose function that computes the
proportion of values from a sample generated
via Monte Carlo methods that are above or below
a comparison point. Can be used, for example,
to estimate a posterior p-value from MCMC
samples.
}
\examples{
set.seed( 831 ) # For reproducibility
x <- rnorm( 1000, mean = 1 )
# Two-sided comparison against 0
p_value.monte_carlo( x )
# One-sided comparison against 0
p_value.monte_carlo( x, alternative = 'greater' )
p_value.monte_carlo( x, alternative = 'less' )
# Non-zero comparison
p_value.monte_carlo( x, comparison = 1 )
p_value.monte_carlo( x, comparison = 1, alternative = 'greater' )
p_value.monte_carlo( x, comparison = 1, alternative = 'less' )

}
