% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S04-predict_cell_means.R
\name{predict_cell_means}
\alias{predict_cell_means}
\title{Predict Cell Means Based on Grouping Factors}
\usage{
predict_cell_means(est, grp = NULL, dm = NULL, lwr = NULL, ...)
}
\arguments{
\item{est}{Output from the \code{\link{fit_with_brm}}
function.}

\item{grp}{A vector with the column names for the
grouping factors.}

\item{dm}{An optional data frame with the grouping
factors and the weights for the population-level
effects needed to compute the cell means.}

\item{lwr}{A list with 1) the column names for the
grouping terms used in a mixed effects model, and
2) any additional variables not part of the
population-level effects.}

\item{...}{Additional parameters for
\code{\link[brms]{predict.brmsfit}}.}
}
\value{
Output from the \code{\link[brms]{predict.brmsfit}}
function.
}
\description{
A convenience function to generate model predictions
over specified grouping factors for population-level
effects.
}
\examples{
# Load in example for linear mixed effects model
data("sleepstudy", package = "lme4")
df <- sleepstudy
df$LC.ZS.Days <- scale( df$Days )[,1]
# Fit model with linear contrast for days
# with subject-level intercepts and
# slopes
alg <- algorithm_settings( desired_iter = 5000,
                           cores = 1,
                           adapt_delta = .8,
                           rng_seed = 111 )
est <- fit_with_brm(
  df, Reaction ~ LC.ZS.Days + (1 + LC.ZS.Days|Subject),
  algorithm = alg
)

# Predict averages per day (no subject-level
# uncertainty)
tst <- predict_cell_means(
  est, grp = 'Days'
)
print( round( tst, 1 ) )

# Predict averages per day (with subject-level
# uncertainty)
tst <- predict_cell_means(
  est, grp = 'Days', lwr = list( 'ID', 'LC.ZS.Days' )
)
print( round( tst, 1 ) )

}
