% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions_for_estimation.R
\name{pull_param}
\alias{pull_param}
\title{Title}
\usage{
pull_param(
  x,
  type = "b_",
  include = NULL,
  exclude = NULL,
  special = NULL,
  and = c(TRUE, FALSE)
)
}
\arguments{
\item{type}{...}

\item{include}{...}

\item{exclude}{...}

\item{special}{...}

\item{and}{...}

\item{prm}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
