% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S01-General_purpose_functions_for_estimation.R
\name{summary.monte_carlo}
\alias{summary.monte_carlo}
\title{Title}
\usage{
\method{summary}{monte_carlo}(x, form = NULL, tidy = TRUE)
}
\arguments{
\item{x}{...}

\item{form}{...}

\item{tidy}{...}
}
\value{
Output.
}
\description{
Description.
}
\details{

}
\examples{
# Examples

}
