% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S03-beta_binomial_regression_setup.R
\name{beta_binomial_regression_setup}
\alias{beta_binomial_regression_setup}
\title{Title}
\usage{
beta_binomial_regression_setup(global = FALSE)
}
\arguments{
\item{global}{Logical; if \code{TRUE} uses
the \code{<<-} assignment to add all
functions and terms to the global environment.}
}
\value{
Output.
}
\description{
Description.
}
\details{
Forthcoming.
}
\examples{
# Examples

}
